Merge everything into one step
This means we have one method that generates, compiles, and executes one
context in a single phase. This function is independent and thread safe,
meaning they could be potentially executed in parallel.

Additionally, it contains the start of some renaming for better clarity.

Only the Python executor has been tested in this commit.