Properly handle invalid utf8 in output streams
The processes that run the generated executable are executed using
subprocess.run(). By default, the output streams (e.g. stdout, stderr)
are assumed to be in utf-8, using the "strict" mode. This means invalid
utf8 in the output streams caused the judge to crash.

The fix is to enable the "backslashreplace" mode, which replaces invalid
sequences by a backslash-escaped version.

See https://docs.python.org/3/library/io.html#io.TextIOWrapper.

Also adds a test for this issue.