Support explicit static functions
Explicitly static functions require static function support from the
programming language, implying OOP support. Using a static function in
a test suite thus causes the test suite to only be usable in OOP
languages (thus not Bash, C or Haskell).

Note that this has no impact on the existing support for static
functions: a top-level function is still translated to a static function
in those languages that require it (i.e. Java).

Additionally, the convention in the DSL that an identifier that starts
with an uppercase letter is a class is now also applied to function
calls: calling a function on an identifier starting with an uppercase
letter is interpreted as being a static call (e.g. Class.method()).

This change is mostly backwards compatible, and there is no known use of
the DSL besides our own.