Update default interpretation of string/objects
YAML strings and objects are now interpreted as literal YAML values by
default.

We introduce `!expression` and `!oracle` to "cast" strings and objects
to expressions and oracles respectively. The list of arguments in an
oracle uses the same logic: YAML types by default, expressions when
tagged as such.

Using the same logic for arguments is not that obvious: we could also
force the use of the expression syntax, as we do with actual expressions
and statements. However, it felt more natural to be consistent with the
return value (which is in the same YAML object) than with the
expressions and statements.

This is basically a reversal of the previous behaviour, where strings
and objects needed a `!v` tag. This tag has now been removed.