Improve JavaScript handling of Map/Object
Switch to Map by default and introduce two new advanced types to make
the distinction between Map/Object when needed. The new types are
"dictionary" (since "map" is already in use) for the datastructure and
"object" for the JavaScript-specific "Object" type, which no other
language natively supports at the moment.

All existing languages have been updated to support the "dictionary"
advanced type if they supported the basic "map" type, while the advanced
type for "object" is reduced, except in JavaScript, which supports both.

As part of this, the way to indicate type limits on certain elements
of collections has been reworked to be more flexible. It is now possible
to specify limits for all types separately. At the moment, we still only
recognize limits on map keys and set elements, so nothing changes there.