Support "raw" identifiers
Using __X__ as identifiers for those that need to be kept as is, e.g.
not conventionalized is a mostly nice way of doing it.

One thing that still needs to consider is how to do constructors.
Since we use the convention that a constructor needs to start with
a capital, which no longer works.

One solution, e.g. new__X__, is ugly and difficult to implement. Another
would be to add e.g. a new function, to allow writing

constructor(__X__, *arguments)

This works cleanly, but is also not that nice.